hw/imc: Do scoms on the secondary core in fused core mode for core-imc counters
Core IMC hardware mandates initiating of three scoms to enable or
disable of the Core IMC engine. This has to be done for each core,
and is done in the opal_imc_counters_init().

In fused core mode, the scom has to be done explicitly for the secondary
core to enable or disable core IMC engine in that core.

Do the scom for the secondary core, by calculating the core id from
the incoming cpu id.

This patch is based on the series,
        Initial fused-core support for POWER9 [v5]:
	https://lists.ozlabs.org/pipermail/skiboot/2020-April/016778.html

Signed-off-by: Anju T Sudhakar <anju@linux.vnet.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>