Don't warn about stack size on host binaries
I'm hitting a stack size warning when building pflash:

common/arch_flash_powerpc.c: In function ‘get_dev_mtd.constprop’:
common/arch_flash_powerpc.c:177:1: error: the frame size of 8240
    bytes is larger than 2048 bytes [-Werror=frame-larger-than=]

That function has 2 PATH_MAX strings, each of which will use up 4kB of
stack.

We've tried to work around the issue of stack size warnings on host
binaries in a few places, with limited success. This patch removes the
check completely instead. We need to modify the HOSTCFLAGS variable
assignment to be immediate for this to work.

Signed-off-by: Anton Blanchard <anton@ozlabs.org>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>