hw/chiptod: Retry the sync procedure on failure
The chiptod sync will sometimes fail and then sync successfully after a
retry.  So, try an arbitrary 10 numbers of times before we either
abort() on main procedure fail or disable threads on secondary procedure
fail.  Also, put a message on the log if secondaries fail so we have
evidence in the log when they aren't enabled.

Signed-off-by: Ryan Grimm <grimm@linux.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>