AWAN simulator support for P10
This patch enables Skiboot to initialize and Linux to boot to user space
on the AWAN core and chip models.

We need the distinction between core and chip models because the core
models do not have an XSCOM unit, CHIPTOD, nor RNG.  The chip
model does have them and they work.

So, add a device_type property to the awan node to distinguish core from
chip.  Sample DTS are provided for the core and chip models in
external/awan.

Just like Mambo, we need to return in slw_init before trying to
initialize SLW.  Without an XSCOM unit in the device tree for the core
model, the SLW code path eventually fails an assert due to lack of
chips.

This commit defines a QUIRK_AWAN where previously Mambo used
QUIRK_MAMBO_CALLOUTS so now Mambo and AWAN core both work.

Also, fix up chip quirks so the core model and chip model boot and
initialize the appropriate units.

Disable sreset and power management in a couple spots because the chip
model does not support stop with EC=1 and enter_p9_pm_state spins in the
branch-to-self after stop.

Provide an external/awan/README.md with a high-level view of booting in
the environment.

Signed-off-by: Ryan Grimm <grimm@linux.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>