phb4: Advertise a trigger page when StoreEOI is activated
Interrupt trigger on a PHB ESB page is supported and it is used by the
OS to synthesize an interrupt when using LoadEOI. However, when
StoreEOI is activated, there is no need to retrigger and the PHB
sources are not flagged with XIVE_SRC_TRIGGER_PAGE which defines the
source trigger page to NULL at the OS level.

When a PHB is passthrough in a KVM guest, the guest VMA associated
with the source ESBs is updated with the PHB ESB pages on page
fault. But, if the guest is running in P9 compat, it will use LoadEOI
and possibly the trigger page. If StoreEOI is activated on the host,
which is the default for P10, the ESB page fault handler will crash
the guest because of a NULL trigger page.

Since the PHB ESBs support stores to trigger the interrupt, add the
XIVE_SRC_TRIGGER_PAGE flag to return a trigger page to the OS.

Signed-off-by: Cédric Le Goater <clg@kaod.org>