core: Add debugfs framework
This is a very simplistic framework adding a read/write interface to
opal drivers. Drivers should define a set of 'opal_debug' handlers,
each with custom read/write operations. The device tree is then
populated with nodes such as :

	debug@0 {
		ibm,chip-id = <0x00>;
		label = "xive-ivt";
		compatible = "ibm,opal-debug";
		reg = <0x00>;
		phandle = <0x805e>;
	};

which can be freely extended by the driver if needed.

The Linux driver can choose to expose the contents of each 'opal_debug'
handler through debugfs using the OPAL_DEBUG_READ call. It is relatively
easy to implement a simple command interface with the OPAL_DEBUG_WRITE
call, to set some values or reset some counters.

Signed-off-by: Cédric Le Goater <clg@kaod.org>