secvar/edk2: store timestamp variable in protected storage
Each signed variable update contains a timestamp -- this timestamp is checked
against the previous timestamp seen for that particular variable (if any), and
the update is rejected if the timestamp is not a later time than the previous.

This timestamp check is intended to prevent re-use of signed update files.
Currently, the code stores the timestamps in the TS variable, which is then
stored in regular variable storage (typically PNOR). This patch promotes the
variable to "protected storage" (typically TPM NV), so avoid this variable
being accidentally cleared.

This change should only come into effect when either:
 - initializing secvar for the first time (i.e. first boot, or
    after a key-clear-request)
 - processing any variable update

Systems that already have a TS variable in PNOR will not be affected until
either of the above actions are taken.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>
Tested-by: Nick Child <nick.child@ibm.com>
Reviewed-by: Daniel Axtens <dja@axtens.net>
Signed-off-by: Cédric Le Goater <clg@kaod.org>
(cherry picked from commit 59a247e7f4e9df2521ebb53cdc47aaa34c225fea)
Signed-off-by: Cédric Le Goater <clg@kaod.org>