CI: Rename ubuntu-latest to ubuntu-rolling
In the docker world, ubuntu-latest is the latest LTS release, 20.04 as
of this writing. ubuntu-rolling is the latest (non-devel) release,
which is 21.10 as of this writing. So rename our CI files accordingly
to avoid confusion.

Also ubuntu 21.10 ships with a recent enough qemu-system-ppc package
so we can now run a simple qemu boot test for powernv. The Docker file
fetches a kernel image from the op-build repo on github.

Signed-off-by: Frederic Barrat <fbarrat@linux.ibm.com>
Signed-off-by: Cédric Le Goater <clg@kaod.org>