github: Add workflow to build and deploy docs
This builds the gcov coverage report and the skiboot docs on the Github
hosted runner, without using a container.

Remove the CROSS= definition from build-docs.sh to allow the makefile
detect which cross compiler is installed, instead of hardcoding the
non-le variant.

This is a simplification from the previous docs build, which used the
containers. However the containers have since been re-worked and no
longer leave the build artifacts on the host system.

The github action used for deploying seems to be the most commonly used:

 https://github.com/JamesIves/github-pages-deploy-action

Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Cédric Le Goater <clg@kaod.org>